test/locust_test/requirements.txt
locust==2.34.0
python-dotenv==1.0.1